Updated actionspooc + added proper speed-ups
Cloakers will only speed up for other peers (now including the host, instead of just clients), but ONLY if they're staying behind compared to what the victim is syncing. It will never affect the victim like it used to, so you can now safely look away when running from them as a client.

The rest of the changes are just polishing, except for one case for non-local attacks where Cloakers wouldn't properly wait if there's no nav_points to go to when the action is started.