Detach corpse units when enabling corpse disposal
Aka prevent the host's corpse limit from nuking clients if a civilian/enemy is killed before the corpse disposal system is enabled, and said civilian/enemy hasn't yet despawned for the host. Yes, this is another cause of the infinite smoke/nothing despawns bug that plagues clients.

To explain a bit, the game normally only detaches dead civilians/enemies from the network when the corpse disposal system isn't enabled, which doesn't necessarily mean full stealth (when you see the detection UI), just that enemies are not on their way yet. You usually see a hint message mentioning that the police were alerted/are on their way, which means the disposal system is active. If no message pops up, and enemies are spawning or assaults are starting (keep in mind that some maps can delay both of these), the system is also enabled.

The thing is, clients don't seem to be able to dispose of units attached to the network (I can't really verify the actual code behind this, but it would make sense if it's that way) and instead wait for the host to send the order.

A good example of this issue is Cook Off (the standalone heist), where if you kill a gangster or rush the house too quickly and the cooks get killed before the cops are called (either by one of the gangsters or automatically through an unnecessarily delayed event, still don't know why it's not done like in Rats day 1), and the host has corpse settings set to ultra, clients are screwed until the ridiculously high limit is reached.