Experimental queued tasks handling rework
With the goal of optimizing the system again.

Basically sort tasks according to their timer and only check the first one (instead of the entire table). If the timer has passed, execute it and check the next one and do the same. If a task has no timer, they're handled through another table.